Implemented basic functions for game.
The cell just holds state information about the cell. The cell can also
return a list of who its neighbor cells are.
The board has a map of the currently alive cells.
The board uses a future map to evaluate which cells survive and which
dead cells come alive.
Since the board uses a map instead of an array it can grow and move as
needed.